Scroll to proper position in annotations
If there is a selection instead of a single line, use the first line in the
selection range.

Takes offset from the navigation box into account and scrolls
to the first row of the commit (scrolling to the **selected** row is
pretty useless, as there is no author or revision info if you are in
the middle of the document).