Fix for the diff action in the commit window
Starting with mate 2.5 it will not read from standard input if the TM_DOCUMENT_UUID environment variable is set. This is unfortunate for commands launched by TextMate (where the variable is set) that use mate to pass data to TextMate (via mate’s stdin).

This fix is provisional until a new version of mate gets released.