When amending auto populate the commit window with last commit message
Since we auto populate the textview with the last commit message, we eliminate the need to include the "-- update commit message --" item. This item was used to inform the bundle if the user intended to update the commit message (check) or just use the previous message (uncheck).

So now when amending a commit, the commit window will show the last commit message in the textview and a list of potential files only if necessary (i.e., the table view will be empty in situations where the user wanted just to edit the commit message). This behavior is more consistent with the way amending a commit works when running "git commit --amend" from the command line.

Also, we now check to see if there is nothing to amend (i.e., it is the initial commit).