Add API to suppress/unsuppress (html) output
This is useful when we would like to discard and not show any potential (html) output that may have been generated up to that point (e.g., canceling an action).

Note that all output will be suppressed until "unsupress_output" is called. I.e., calling "flush" will not show the current output buffer.