Allow a commit summary of up to 72 characters before marking it invalid
From ‘git help commit’ the recommendation is simply to make it ‘less than 50 characters’ which we still use as the limit for marking it as deprecated.

I think GitHub used to truncate at 65 characters but currently they truncate at 72 which is a commonly used value for hard wrapping, so it makes sense for us to use this as well.